Move "test examples" to examples
